requirements.txt
openai==1.3.9
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.0
scikit-learn==1.3.0
sentence-transformers==2.2.2
tqdm==4.66.1
requests==2.31.0
transformers==4.29.2
tokenizers==0.13.3
huggingface-hub==0.14.1
torch==2.0.1
pathlib==1.0.1
urllib3==2.0.7
faiss-cpu==1.11.0
